新增功能: 新增會員WEB API現在多新增支援舊會員轉檔

+++ LoveMeHandMake2/Models/ApiModels/DepositResultApiModel.cs
@@ -18,7 +18,7 @@ namespace LoveMeHandMake2.Models.ApiModels
         public int DepositRewardPoint { get; set; }
 
         [Display(Name = "总新增点数")]
-        public int TotalPoint { get; set; }
+        public double TotalPoint { get; set; }
 
         [Display(Name = "每点平均成本")]
         public double AvgPointCost { get; set; }
